requirements.txt
bcrypt==4.1.1
click==8.1.7
iniconfig==2.0.0
packaging==23.2
pluggy==1.3.0
PyJWT==2.8.0
pytest==7.4.3
python-dotenv==1.0.0
SQLAlchemy==2.0.23
typing_extensions==4.8.0
